2017.4.16f1

+++ CGIncludes/UnityStandardBRDF.cginc
@@ -80,12 +80,12 @@ inline half4 Pow5 (half4 x)
 
 inline half3 FresnelTerm (half3 F0, half cosA)
 {
-	half t = Pow5 (1 - cosA);	// ala Schlick interpoliation
+	half t = Pow5 (1 - cosA);   // ala Schlick interpoliation
 	return F0 + (1-F0) * t;
 }
 inline half3 FresnelLerp (half3 F0, half3 F90, half cosA)
 {
-	half t = Pow5 (1 - cosA);	// ala Schlick interpoliation
+	half t = Pow5 (1 - cosA);   // ala Schlick interpoliation
 	return lerp (F0, F90, t);
 }
 // approximage Schlick with ^4 instead of ^5
@@ -100,8 +100,8 @@ half DisneyDiffuse(half NdotV, half NdotL, half LdotH, half perceptualRoughness)
 {
 	half fd90 = 0.5 + 2 * LdotH * LdotH * perceptualRoughness;
 	// Two schlick fresnel term
-	half lightScatter	= (1 + (fd90 - 1) * Pow5(1 - NdotL));
-	half viewScatter	= (1 + (fd90 - 1) * Pow5(1 - NdotV));
+	half lightScatter   = (1 + (fd90 - 1) * Pow5(1 - NdotL));
+	half viewScatter    = (1 + (fd90 - 1) * Pow5(1 - NdotV));
 
 	return lightScatter * viewScatter;
 }
@@ -127,29 +127,29 @@ inline half SmithBeckmannVisibilityTerm (half NdotL, half NdotV, half roughness)
 }
 
 // Ref: http://jcgt.org/published/0003/02/03/paper.pdf
-inline half SmithJointGGXVisibilityTerm (half NdotL, half NdotV, half roughness)
+inline float SmithJointGGXVisibilityTerm (float NdotL, float NdotV, float roughness)
 {
 #if 0
 	// Original formulation:
-	// lambda_v	= (-1 + sqrt(a2 * (1 - NdotL2) / NdotL2 + 1)) * 0.5f;
-	// lambda_l	= (-1 + sqrt(a2 * (1 - NdotV2) / NdotV2 + 1)) * 0.5f;
-	// G			= 1 / (1 + lambda_v + lambda_l);
+	// lambda_v    = (-1 + sqrt(a2 * (1 - NdotL2) / NdotL2 + 1)) * 0.5f;
+	// lambda_l    = (-1 + sqrt(a2 * (1 - NdotV2) / NdotV2 + 1)) * 0.5f;
+	// G           = 1 / (1 + lambda_v + lambda_l);
 
 	// Reorder code to be more optimal
-	half a		= roughness;
-	half a2		= a * a;
+	half a          = roughness;
+	half a2         = a * a;
 
-	half lambdaV	= NdotL * sqrt((-NdotV * a2 + NdotV) * NdotV + a2);
-	half lambdaL	= NdotV * sqrt((-NdotL * a2 + NdotL) * NdotL + a2);
+	half lambdaV    = NdotL * sqrt((-NdotV * a2 + NdotV) * NdotV + a2);
+	half lambdaL    = NdotV * sqrt((-NdotL * a2 + NdotL) * NdotL + a2);
 
 	// Simplify visibility term: (2.0f * NdotL * NdotV) / ((4.0f * NdotL * NdotV) * (lambda_v + lambda_l + 1e-5f));
 	return 0.5f / (lambdaV + lambdaL + 1e-5f); // This function is not intended to be running on Mobile,
 												// therefore epsilon is smaller than can be represented by half
 #else
 	// Approximation of the above formulation (simplify the sqrt, not mathematically correct but close enough)
-	half a = roughness;
-	half lambdaV = NdotL * (NdotV * (1 - a) + a);
-	half lambdaL = NdotV * (NdotL * (1 - a) + a);
+	float a = roughness;
+	float lambdaV = NdotL * (NdotV * (1 - a) + a);
+	float lambdaL = NdotV * (NdotL * (1 - a) + a);
 
 	return 0.5f / (lambdaV + lambdaL + 1e-5f);
 #endif
@@ -165,10 +165,10 @@ inline float GGXTerm (float NdotH, float roughness)
 
 inline half PerceptualRoughnessToSpecPower (half perceptualRoughness)
 {
-	half m = PerceptualRoughnessToRoughness(perceptualRoughness);	// m is the true academic roughness.
+	half m = PerceptualRoughnessToRoughness(perceptualRoughness);   // m is the true academic roughness.
 	half sq = max(1e-4f, m*m);
-	half n = (2.0 / sq) - 2.0;						// https://dl.dropboxusercontent.com/u/55891920/papers/mm_brdf.pdf
-	n = max(n, 1e-4f);								// prevent possible cases of pow(0,0), which could happen when roughness is 1.0 and NdotH is zero
+	half n = (2.0 / sq) - 2.0;                          // https://dl.dropboxusercontent.com/u/55891920/papers/mm_brdf.pdf
+	n = max(n, 1e-4f);                                  // prevent possible cases of pow(0,0), which could happen when roughness is 1.0 and NdotH is zero
 	return n;
 }
 
@@ -249,12 +249,12 @@ half4 BRDF1_Unity_PBS (half3 diffColor, half3 specColor, half oneMinusReflectivi
 	// A re-normalization should be applied here but as the shift is small we don't do it to save ALU.
 	//normal = normalize(normal);
 
-	half nv = saturate(dot(normal, viewDir)); // TODO: this saturate should no be necessary here
+	float nv = saturate(dot(normal, viewDir)); // TODO: this saturate should no be necessary here
 #else
-	half nv = abs(dot(normal, viewDir));	// This abs allow to limit artifact
+	half nv = abs(dot(normal, viewDir));    // This abs allow to limit artifact
 #endif
 
-	half nl = saturate(dot(normal, light.dir));
+	float nl = saturate(dot(normal, light.dir));
 	float nh = saturate(dot(normal, halfDir));
 
 	half lv = saturate(dot(light.dir, viewDir));
@@ -271,7 +271,7 @@ half4 BRDF1_Unity_PBS (half3 diffColor, half3 specColor, half oneMinusReflectivi
 #if UNITY_BRDF_GGX
 	// GGX with roughtness to 0 would mean no specular at all, using max(roughness, 0.002) here to match HDrenderloop roughtness remapping.
 	roughness = max(roughness, 0.002);
-	half V = SmithJointGGXVisibilityTerm (nl, nv, roughness);
+	float V = SmithJointGGXVisibilityTerm (nl, nv, roughness);
 	float D = GGXTerm (nh, roughness);
 #else
 	// Legacy
@@ -279,11 +279,11 @@ half4 BRDF1_Unity_PBS (half3 diffColor, half3 specColor, half oneMinusReflectivi
 	half D = NDFBlinnPhongNormalizedTerm (nh, PerceptualRoughnessToSpecPower(perceptualRoughness));
 #endif
 
-	half specularTerm = V*D * UNITY_PI; // Torrance-Sparrow model, Fresnel is applied later
+	float specularTerm = V*D * UNITY_PI; // Torrance-Sparrow model, Fresnel is applied later
 
-#	ifdef UNITY_COLORSPACE_GAMMA
+	#ifdef UNITY_COLORSPACE_GAMMA
 		specularTerm = sqrt(max(1e-4h, specularTerm));
-#	endif
+	#endif
 
 	// specularTerm * nl can be NaN on Metal in some cases, use max() to make sure it's a sane value
 	specularTerm = max(0, specularTerm * nl);
@@ -293,17 +293,17 @@ half4 BRDF1_Unity_PBS (half3 diffColor, half3 specColor, half oneMinusReflectivi
 
 	// surfaceReduction = Int D(NdotH) * NdotH * Id(NdotL>0) dH = 1/(roughness^2+1)
 	half surfaceReduction;
-#	ifdef UNITY_COLORSPACE_GAMMA
-		surfaceReduction = 1.0-0.28*roughness*perceptualRoughness;	// 1-0.28*x^3 as approximation for (1/(x^4+1))^(1/2.2) on the domain [0;1]
-#	else
-		surfaceReduction = 1.0 / (roughness*roughness + 1.0);			// fade \in [0.5;1]
-#	endif
+	#ifdef UNITY_COLORSPACE_GAMMA
+		surfaceReduction = 1.0-0.28*roughness*perceptualRoughness;      // 1-0.28*x^3 as approximation for (1/(x^4+1))^(1/2.2) on the domain [0;1]
+	#else
+		surfaceReduction = 1.0 / (roughness*roughness + 1.0);           // fade \in [0.5;1]
+	#endif
 
 	// To provide true Lambert lighting, we need to be able to kill specular completely.
 	specularTerm *= any(specColor) ? 1.0 : 0.0;
 
 	half grazingTerm = saturate(smoothness + (1-oneMinusReflectivity));
-	half3 color =	diffColor * (gi.diffuse + light.color * diffuseTerm)
+	half3 color = diffColor * (gi.diffuse + light.color * diffuseTerm)
 					+ specularTerm * light.color * FresnelTerm (specColor, lh)
 					+ surfaceReduction * gi.specular * FresnelLerp (specColor, grazingTerm, nv);
 
@@ -387,7 +387,7 @@ half4 BRDF2_Unity_PBS (half3 diffColor, half3 specColor, half oneMinusReflectivi
 	// surfaceReduction = Int D(NdotH) * NdotH * Id(NdotL>0) dH = 1/(realRoughness^2+1)
 
 	// 1-0.28*x^3 as approximation for (1/(x^4+1))^(1/2.2) on the domain [0;1]
-	// 1-x^3*(0.6-0.08*x)	approximation for 1/(x^4+1)
+	// 1-x^3*(0.6-0.08*x)   approximation for 1/(x^4+1)
 #ifdef UNITY_COLORSPACE_GAMMA
 	half surfaceReduction = 0.28;
 #else
@@ -397,7 +397,7 @@ half4 BRDF2_Unity_PBS (half3 diffColor, half3 specColor, half oneMinusReflectivi
 	surfaceReduction = 1.0 - roughness*perceptualRoughness*surfaceReduction;
 
 	half grazingTerm = saturate(smoothness + (1-oneMinusReflectivity));
-	half3 color =	(diffColor + specularTerm * specColor) * light.color * nl
+	half3 color = (diffColor + specularTerm * specColor) * light.color * nl
 					+ gi.diffuse * diffColor
 					+ surfaceReduction * gi.specular * FresnelLerpFast (specColor, grazingTerm, nv);
